backend/risk_engine/risk_engine/rag_agent_hf/requirements.txt
fastapi
uvicorn[standard]
pydantic
python-multipart
tqdm
transformers>=4.35
sentence-transformers
datasets
faiss-cpu
smolagents
huggingface-hub
pdfminer.six
python-docx
langchain
numpy
scikit-learn
torch
